Added a rerender to update the regiter token status
