+++ client/src/pages/HiringManagement/HiringManagement.jsx
@@ -1,8 +1,9 @@
-import { Box, Tab, Tabs } from '@mui/material';
-import React from 'react';
+import { Box, Tab, Tabs, CircularProgress } from '@mui/material';
+import React, { useEffect, useState } from 'react';
 import ApplicationTables from './ViewOnboardingApplications';
 import SendLink from './SendRegistration';
 import NewUserList from './NewUserList';
+import { get } from '../../services/api';  // Import your API utility
 
 function a11yProps(index) {
     return {
@@ -12,11 +13,41 @@ function a11yProps(index) {
 }
 
 const HiringManagement = () => {
-    const [value, setValue] = React.useState(0);
+    const [value, setValue] = useState(0);
+    const [users, setUsers] = useState([]);
+    const [isLoading, setIsLoading] = useState(true);
+    const [error, setError] = useState(null);
 
     const handleChange = (event, newValue) => {
         setValue(newValue);
     };
+
+    // Function to fetch users
+    const fetchUsers = async () => {
+        setIsLoading(true);
+        try {
+            const data = await get('/user/all-users');
+            setUsers(data);
+        } catch (err) {
+            setError('Error fetching users.');
+            console.error(err);
+        } finally {
+            setIsLoading(false);
+        }
+    };
+
+    // Fetch users when the component mounts
+    useEffect(() => {
+        fetchUsers();
+    }, []);
+
+    if (isLoading) {
+        return <CircularProgress />;
+    }
+
+    if (error) {
+        return <Typography color="error">{error}</Typography>;
+    }
 
     return (
         <Box>
@@ -26,13 +57,13 @@ const HiringManagement = () => {
             </Tabs>
             <Box
                 index={0}
-                role='tabpane0'
+                role='tabpanel'
                 hidden={value !== 0}
                 id={'simple-tabpanel-0'}
                 aria-labelledby={`simple-tab-0`}
             >
-                <SendLink />
-                <NewUserList />
+                <SendLink setUsers={setUsers} fetchUsers={fetchUsers} />
+                <NewUserList users={users} />
             </Box>
             <ApplicationTables index={1} value={value} />
         </Box>
